collector/requirements.txt
flask
pydantic
requests
python-dotenv
